+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.1.0
+
+- Added `--truncate` option to truncate the output of the search results
+- Added the ability to exclude lines from the search results
+
 ## v1.0.0
 
 - Removed default values for `--exclude` and `--regex` options